Move EuropeanVAT class to different file
